+++ plugins/ssr/ssripats.ml
@@ -27,14 +27,30 @@ module IpatMachine : sig
   val main : ?eqtac:unit tactic -> first_case_is_dispatch:bool ->
         ssripats -> unit tactic
 
+
+  val tclSEED_SUBGOALS : Names.Name.t list array -> unit tactic -> unit tactic
+
 end = struct (* {{{ *)
 
 module State : sig
+
+  type delayed_gen = {
+    tmp_id : Id.t;    (* Temporary name *)
+    orig_name : Name.t   (* Old name *)
+  }
 
   (* to_clear API *)
   val isCLR_PUSH    : Id.t -> unit tactic
   val isCLR_PUSHL   : Id.t list -> unit tactic
   val isCLR_CONSUME : unit tactic
+
+  (* to_generalize API *)
+  val isGEN_PUSH    : delayed_gen -> unit tactic
+  val isGEN_CONSUME : unit tactic
+
+  (* name_seed API *)
+  val isNSEED_SET : Names.Name.t list -> unit tactic
+  val isNSEED_CONSUME : (Names.Name.t list option -> unit tactic) -> unit tactic
 
   (* Some data may expire *)
   val isTICK : ssripat -> unit tactic
@@ -48,10 +64,23 @@ type istate = {
   (* Delayed clear *)
   to_clear : Id.t list;
 
+  (* Temporary intros, to be generalized back *)
+  to_generalize : delayed_gen list;
+
+  (* The type of the inductive constructor corresponding to the current proof
+   * branch: name seeds are taken from that in an intro block *)
+  name_seed : Names.Name.t list option;
+
+}
+and delayed_gen = {
+  tmp_id : Id.t;    (* Temporary name *)
+  orig_name : Name.t   (* Old name *)
 }
 
 let empty_state = {
   to_clear = [];
+  to_generalize = [];
+  name_seed = None;
 }
 
 include Ssrcommon.MakeState(struct
@@ -59,28 +88,69 @@ include Ssrcommon.MakeState(struct
   let init = empty_state
 end)
 
+let print_name_seed env sigma = function
+  | None -> Pp.str "-"
+  | Some nl -> Pp.prlist Names.Name.print nl
+
+let print_delayed_gen { tmp_id; orig_name } =
+  Pp.(Id.print tmp_id ++ str"->" ++ Name.print orig_name)
+
 let isPRINT g =
+  let env, sigma = Goal.env g, Goal.sigma g in
   let state = get g in
   Pp.(str"{{ to_clear: " ++
         prlist_with_sep spc Id.print state.to_clear ++ spc () ++
-      str" }}")
+      str"to_generalize: " ++
+        prlist_with_sep spc print_delayed_gen state.to_generalize ++ spc () ++
+      str"name_seed: " ++ print_name_seed env sigma state.name_seed ++ str" }}")
 
 
 let isCLR_PUSH id =
-  tclGET (fun { to_clear = ids } ->
-  tclSET { to_clear = id :: ids })
+  tclGET (fun ({ to_clear = ids } as s) ->
+  tclSET { s with to_clear = id :: ids })
 
 let isCLR_PUSHL more_ids =
-  tclGET (fun { to_clear = ids } ->
-  tclSET { to_clear = more_ids @ ids })
+  tclGET (fun ({ to_clear = ids } as s) ->
+  tclSET { s with to_clear = more_ids @ ids })
 
 let isCLR_CONSUME =
-  tclGET (fun { to_clear = ids } ->
-  tclSET { to_clear = [] } <*>
+  tclGET (fun ({ to_clear = ids } as s) ->
+  tclSET { s with to_clear = [] } <*>
   Tactics.clear ids)
 
 
-let isTICK _ = tclUNIT ()
+let isGEN_PUSH dg =
+  tclGET (fun s ->
+  tclSET { s with to_generalize = dg :: s.to_generalize })
+
+(* generalize `id` as `new_name` *)
+let gen_astac id new_name =
+ let gen = ((None,Some(false,[])),Ssrmatching.cpattern_of_id id) in
+ V82.tactic (Ssrcommon.gentac gen)
+ <*> Ssrcommon.tclRENAME_HD_PROD new_name
+
+(* performs and resets all delayed generalizations *)
+let isGEN_CONSUME =
+  tclGET (fun ({ to_generalize = dgs } as s) ->
+  tclSET { s with to_generalize = [] } <*>
+  Tacticals.New.tclTHENLIST
+    (List.map (fun { tmp_id; orig_name } ->
+       gen_astac tmp_id orig_name) dgs) <*>
+  Tactics.clear (List.map (fun gen -> gen.tmp_id) dgs))
+
+
+let isNSEED_SET ty =
+  tclGET (fun s ->
+  tclSET { s with name_seed = Some ty })
+
+let isNSEED_CONSUME k =
+  tclGET (fun ({ name_seed = x } as s) ->
+  tclSET { s with name_seed = None } <*>
+  k x)
+
+let isTICK = function
+  | IPatSimpl _ | IPatClear _ -> tclUNIT ()
+  | _ -> tclGET (fun s -> tclSET { s with name_seed = None })
 
 end (* }}} *************************************************************** *)
 
@@ -105,18 +175,49 @@ let intro_anon_all = Goal.enter begin fun gl ->
   let sigma = Goal.sigma gl in
   let g = Goal.concl gl in
   let n = nb_assums env sigma g in
-  Tacticals.New.tclDO n Ssrcommon.tclINTRO_ANON
+  Tacticals.New.tclDO n (Ssrcommon.tclINTRO_ANON ())
+end
+
+(*** [=> >*] **************************************************************)
+(** [nb_deps_assums] returns the number of dependent premises *)
+let rec nb_deps_assums cur env sigma t =
+  let t' = Reductionops.whd_allnolet env sigma t in
+  match EConstr.kind sigma t' with
+  | Constr.Prod(name,ty,body) ->
+     if EConstr.Vars.noccurn sigma 1 body &&
+        not (Typeclasses.is_class_type sigma ty) then cur
+     else nb_deps_assums (cur+1) env sigma body
+  | Constr.LetIn(name,ty,t1,t2) ->
+     nb_deps_assums (cur+1) env sigma t2
+  | Constr.Cast(t,_,_) ->
+     nb_deps_assums cur env sigma t
+  | _ -> cur
+let nb_deps_assums = nb_deps_assums 0
+
+let intro_anon_deps = Goal.enter begin fun gl ->
+  let env = Goal.env gl in
+  let sigma = Goal.sigma gl in
+  let g = Goal.concl gl in
+  let n = nb_deps_assums env sigma g in
+  Tacticals.New.tclDO n (Ssrcommon.tclINTRO_ANON ())
 end
 
 (** [intro_drop] behaves like [intro_anon] but registers the id of the
     introduced assumption for a delayed clear. *)
 let intro_drop =
-  Ssrcommon.tclINTRO ~id:None
+  Ssrcommon.tclINTRO ~id:Ssrcommon.Anon
     ~conclusion:(fun ~orig_name:_ ~new_name -> isCLR_PUSH new_name)
+
+(** [intro_temp] behaves like [intro_anon] but registers the id of the
+    introduced assumption for a regeneralization. *)
+let intro_anon_temp =
+  Ssrcommon.tclINTRO ~id:Ssrcommon.Anon
+    ~conclusion:(fun ~orig_name ~new_name ->
+      isGEN_PUSH { tmp_id = new_name; orig_name })
 
 (** [intro_end] performs the actions that have been delayed. *)
 let intro_end =
-  Ssrcommon.tcl0G (isCLR_CONSUME)
+  Ssrcommon.tcl0G ~default:() (isCLR_CONSUME <*> isGEN_CONSUME)
 
 (** [=> _] *****************************************************************)
 let intro_clear ids =
@@ -138,6 +239,27 @@ let tacCHECK_HYPS_EXIST hyps = Goal.enter begin fun gl ->
 end
 
 (** [=> []] *****************************************************************)
+
+(* calls t1 then t2 on each subgoal passing to t2 the index of the current
+ * subgoal (starting from 0) as well as the number of subgoals *)
+let tclTHENin t1 t2 =
+  tclUNIT () >>= begin fun () -> let i = ref (-1) in
+  t1 <*> numgoals >>= fun n ->
+  Goal.enter begin fun g -> incr i; t2 !i n end
+end
+
+(* Attaches one element of `seeds` to each of the last k goals generated by
+`tac`, where k is the size of `seeds` *)
+let tclSEED_SUBGOALS seeds tac =
+  tclTHENin tac (fun i n ->
+          Ssrprinters.ppdebug (lazy Pp.(str"seeding"));
+      (* eg [case: (H _ : nat)] generates 3 goals:
+         - 1 for _
+         - 2 for the nat constructors *)
+    let extra_goals = n - Array.length seeds in
+    if i < extra_goals then tclUNIT ()
+    else isNSEED_SET seeds.(i - extra_goals))
+
 let tac_case t =
   Goal.enter begin fun _ ->
     Ssrcommon.tacTYPEOF t >>= fun ty ->
@@ -145,10 +267,36 @@ let tac_case t =
     if is_inj then
       V82.tactic ~nf_evars:false (Ssrelim.perform_injection t)
     else
-      Ssrelim.ssrscasetac t
+      Goal.enter begin fun g ->
+         (Ssrelim.casetac t (fun ?seed k ->
+           match seed with
+           | None -> k
+           | Some seed -> tclSEED_SUBGOALS seed k))
+      end
 end
 
-(***[=> [: id]] ************************************************************)
+(** [=> [^ seed ]] *********************************************************)
+let tac_intro_seed interp_ipats fix = Goal.enter begin fun gl ->
+  isNSEED_CONSUME begin fun seeds ->
+    let seeds =
+      Ssrcommon.option_assert_get seeds Pp.(str"tac_intro_seed: no seed") in
+    let ipats = List.map (function
+       | Anonymous ->
+           let s = match fix with
+             | Prefix id ->  Id.to_string id ^ "?"
+             | SuffixNum n -> "?" ^ string_of_int n
+             | SuffixId id -> "?" ^ Id.to_string id in
+           IPatAnon (One (Some s))
+       | Name id ->
+           let s = match fix with
+             | Prefix fix ->  Id.to_string fix ^ Id.to_string id
+             | SuffixNum n -> Id.to_string id ^ string_of_int n
+             | SuffixId fix -> Id.to_string id ^ Id.to_string fix in
+           IPatId (None, Id.of_string s)) seeds in
+    interp_ipats ipats
+end end
+
+(*** [=> [: id]] ************************************************************)
 [@@@ocaml.warning "-3"]
 let mk_abstract_id =
   let open Coqlib in
@@ -205,95 +353,112 @@ let tclLOG p t =
   end
   <*>
     t p
-  <*>
+  >>= fun ret ->
   Goal.enter begin fun g ->
     Ssrprinters.ppdebug (lazy Pp.(str "done: " ++ isPRINT g));
     tclUNIT ()
   end
+  >>= fun () -> tclUNIT ret
 
-let rec ipat_tac1 ipat : unit tactic =
+let notTAC = tclUNIT false
+
+(* returns true if it was a tactic (eg /ltac:tactic) *)
+let rec ipat_tac1 ipat : bool tactic =
   match ipat with
   | IPatView (clear_if_id,l) ->
-      Ssrview.tclIPAT_VIEWS ~views:l ~clear_if_id
+      Ssrview.tclIPAT_VIEWS
+        ~views:l ~clear_if_id
         ~conclusion:(fun ~to_clear:clr -> intro_clear clr)
 
-  | IPatDispatch(true,[[]]) ->
-      tclUNIT ()
-  | IPatDispatch(_,ipatss) ->
-      tclDISPATCH (List.map ipat_tac ipatss)
+  | IPatDispatch(true, Regular [[]]) ->
+      notTAC
+  | IPatDispatch(_, Regular ipatss) ->
+      tclDISPATCH (List.map ipat_tac ipatss) <*> notTAC
+  | IPatDispatch(_,Block id_block) ->
+      tac_intro_seed ipat_tac id_block <*> notTAC
 
-  | IPatId id -> Ssrcommon.tclINTRO_ID id
+  | IPatId (None, id) -> Ssrcommon.tclINTRO_ID id <*> notTAC
+  | IPatId (Some Dependent, id) ->
+      intro_anon_deps <*> Ssrcommon.tclINTRO_ID id <*> notTAC
 
-  | IPatCase ipatss ->
-     tclIORPAT (Ssrcommon.tclWITHTOP tac_case) ipatss
+  | IPatCase (Block id_block) ->
+      Ssrcommon.tclWITHTOP tac_case <*> tac_intro_seed ipat_tac id_block <*> notTAC
+
+  | IPatCase (Regular ipatss) ->
+     tclIORPAT (Ssrcommon.tclWITHTOP tac_case) ipatss <*> notTAC
   | IPatInj ipatss ->
      tclIORPAT (Ssrcommon.tclWITHTOP
        (fun t -> V82.tactic  ~nf_evars:false (Ssrelim.perform_injection t)))
        ipatss
+     <*> notTAC
 
-  | IPatAnon Drop -> intro_drop
-  | IPatAnon One -> Ssrcommon.tclINTRO_ANON
-  | IPatAnon All -> intro_anon_all
+  | IPatAnon Drop -> intro_drop <*> notTAC
+  | IPatAnon (One seed) -> Ssrcommon.tclINTRO_ANON ?seed () <*> notTAC
+  | IPatAnon All -> intro_anon_all <*> notTAC
+  | IPatAnon Temporary -> intro_anon_temp <*> notTAC
 
-  | IPatNoop -> tclUNIT ()
-  | IPatSimpl Nop -> tclUNIT ()
+  | IPatNoop -> notTAC
+  | IPatSimpl Nop -> notTAC
 
   | IPatClear ids ->
       tacCHECK_HYPS_EXIST ids <*>
-      intro_clear (List.map Ssrcommon.hyp_id ids)
+      intro_clear (List.map Ssrcommon.hyp_id ids) <*>
+      notTAC
 
-  | IPatSimpl (Simpl n) ->
-       V82.tactic ~nf_evars:false (Ssrequality.simpltac (Simpl n))
-  | IPatSimpl (Cut n) ->
-       V82.tactic ~nf_evars:false (Ssrequality.simpltac (Cut n))
-  | IPatSimpl (SimplCut (n,m)) ->
-       V82.tactic ~nf_evars:false (Ssrequality.simpltac (SimplCut (n,m)))
+  | IPatSimpl x ->
+      V82.tactic ~nf_evars:false (Ssrequality.simpltac x) <*> notTAC
 
   | IPatRewrite (occ,dir) ->
      Ssrcommon.tclWITHTOP
-       (fun x -> V82.tactic  ~nf_evars:false (Ssrequality.ipat_rewrite occ dir x))
+       (fun x -> V82.tactic  ~nf_evars:false (Ssrequality.ipat_rewrite occ dir x)) <*> notTAC
 
-  | IPatAbstractVars ids -> tclMK_ABSTRACT_VARS ids
+  | IPatAbstractVars ids -> tclMK_ABSTRACT_VARS ids <*> notTAC
 
-  | IPatTac t -> t
+  | IPatEqGen t -> t <*> notTAC
 
 and ipat_tac pl : unit tactic =
   match pl with
   | [] -> tclUNIT ()
   | pat :: pl ->
-      Ssrcommon.tcl0G (tclLOG pat ipat_tac1) <*>
-      isTICK pat <*>
-      ipat_tac pl
+      Ssrcommon.tcl0G ~default:false (tclLOG pat ipat_tac1) >>= fun was_tac ->
+      isTICK pat (* drops expired seeds *) >>= fun () ->
+      if was_tac then (* exception *)
+        let ip_before, case, ip_after = split_at_first_case pl in
+        let case = ssr_exception true case in
+        let case = option_to_list case in
+        ipat_tac (ip_before @ case @ ip_after)
+      else ipat_tac pl
 
 and tclIORPAT tac = function
   | [[]] -> tac
   | p -> Tacticals.New.tclTHENS tac (List.map ipat_tac p)
 
-let split_at_first_case ipats =
-  let rec loop acc = function
-    | (IPatSimpl _ | IPatClear _) as x :: rest -> loop (x :: acc) rest
-    | IPatCase _ as x :: xs -> CList.rev acc, Some x, xs
-    | pats -> CList.rev acc, None, pats
-  in
-    loop [] ipats
-
-let ssr_exception is_on = function
+and ssr_exception is_on = function
   | Some (IPatCase l) when is_on -> Some (IPatDispatch(true, l))
   | x -> x
 
-let option_to_list = function None -> [] | Some x -> [x]
+and option_to_list = function None -> [] | Some x -> [x]
+
+and split_at_first_case ipats =
+  let rec loop acc = function
+    | (IPatSimpl _ | IPatClear _) as x :: rest -> loop (x :: acc) rest
+    | (IPatCase _ | IPatDispatch _) as x :: xs -> CList.rev acc, Some x, xs
+    | pats -> CList.rev acc, None, pats
+  in
+    loop [] ipats
+;;
 
 (* Simple pass doing {x}/v ->  /v{x} *)
 let elaborate_ipats l =
   let rec elab = function
   | [] -> []
   | (IPatClear _ as p1) :: (IPatView _ as p2) :: rest -> p2 :: p1 :: elab rest
-  | IPatDispatch(s,p) :: rest -> IPatDispatch (s,List.map elab p) :: elab rest
-  | IPatCase p :: rest -> IPatCase (List.map elab p) :: elab rest
+  | IPatDispatch(s, Regular p) :: rest -> IPatDispatch (s, Regular (List.map elab p)) :: elab rest
+  | IPatCase (Regular p) :: rest -> IPatCase (Regular (List.map elab p)) :: elab rest
   | IPatInj p :: rest -> IPatInj (List.map elab p) :: elab rest
-  | (IPatTac _ | IPatId _ | IPatSimpl _ | IPatClear _ |
+  | (IPatEqGen _ | IPatId _ | IPatSimpl _ | IPatClear _ |
      IPatAnon _ | IPatView _ | IPatNoop | IPatRewrite _ |
-     IPatAbstractVars _) as x :: rest -> x :: elab rest
+     IPatAbstractVars _ | IPatDispatch(_, Block _) | IPatCase(Block _)) as x :: rest -> x :: elab rest
   in
     elab l
 
@@ -302,8 +467,8 @@ let main ?eqtac ~first_case_is_dispatch ipats =
   let ip_before, case, ip_after = split_at_first_case ipats in
   let case = ssr_exception first_case_is_dispatch case in
   let case = option_to_list case in
-  let eqtac = option_to_list (Option.map (fun x -> IPatTac x) eqtac) in
-  Ssrcommon.tcl0G (ipat_tac (ip_before @ case @ eqtac @ ip_after) <*> intro_end)
+  let eqtac = option_to_list (Option.map (fun x -> IPatEqGen x) eqtac) in
+  Ssrcommon.tcl0G ~default:() (ipat_tac (ip_before @ case @ eqtac @ ip_after) <*> intro_end)
 
 end (* }}} *)
 
@@ -344,10 +509,11 @@ let mkCoqRefl t c env sigma =
 
 (** Intro patterns processing for elim tactic, in particular when used in
     conjunction with equation generation as in [elim E: x] *)
-let elim_intro_tac ipats ?ist what eqid ssrelim is_rec clr =
+let elim_intro_tac ipats ?seed what eqid ssrelim is_rec clr =
   let intro_eq =
     match eqid with
-    | Some (IPatId ipat) when not is_rec ->
+    | Some (IPatId (Some _, _)) -> assert false (* parser *)
+    | Some (IPatId (None,ipat)) when not is_rec ->
        let rec intro_eq () = Goal.enter begin fun g ->
          let sigma, env, concl = Goal.(sigma g, env g, concl g) in
          match EConstr.kind_of_type sigma concl with
@@ -356,12 +522,12 @@ let elim_intro_tac ipats ?ist what eqid ssrelim is_rec clr =
              | Term.AtomicType (hd, _) when Ssrcommon.is_protect hd env sigma ->
                 V82.tactic ~nf_evars:false Ssrcommon.unprotecttac <*>
                 Ssrcommon.tclINTRO_ID ipat
-             | _ -> Ssrcommon.tclINTRO_ANON <*> intro_eq ()
+             | _ -> Ssrcommon.tclINTRO_ANON () <*> intro_eq ()
              end
          |_ -> Ssrcommon.errorstrm (Pp.str "Too many names in intro pattern")
        end in
        intro_eq ()
-    | Some (IPatId ipat) ->
+    | Some (IPatId (None,ipat)) ->
        let intro_lhs = Goal.enter begin fun g ->
          let sigma = Goal.sigma g in
          let elim_name = match clr, what with
@@ -369,12 +535,9 @@ let elim_intro_tac ipats ?ist what eqid ssrelim is_rec clr =
            | _, `EConstr(_,_,t) when EConstr.isVar sigma t ->
               EConstr.destVar sigma t
            | _ -> Ssrcommon.mk_anon_id "K" (Tacmach.New.pf_ids_of_hyps g) in
-         let elim_name =
-           if Ssrcommon.is_name_in_ipats elim_name ipats then
-             Ssrcommon.mk_anon_id "K" (Tacmach.New.pf_ids_of_hyps g)
-           else elim_name
-         in
-         Ssrcommon.tclINTRO_ID elim_name
+         Tacticals.New.tclFIRST
+           [ Ssrcommon.tclINTRO_ID elim_name
+           ; Ssrcommon.tclINTRO_ANON ~seed:"K" ()]
        end in
        let rec gen_eq_tac () = Goal.enter begin fun g ->
          let sigma, env, concl = Goal.(sigma g, env g, concl g) in
@@ -389,7 +552,7 @@ let elim_intro_tac ipats ?ist what eqid ssrelim is_rec clr =
            | _ -> assert false in
          let case = args.(Array.length args-1) in
          if not(EConstr.Vars.closed0 sigma case)
-         then Ssrcommon.tclINTRO_ANON <*> gen_eq_tac ()
+         then Ssrcommon.tclINTRO_ANON () <*> gen_eq_tac ()
          else
            Ssrcommon.tacTYPEOF case >>= fun case_ty ->
            let open EConstr in
@@ -407,13 +570,14 @@ let elim_intro_tac ipats ?ist what eqid ssrelim is_rec clr =
        intro_lhs <*>
        Ssrcommon.tclINTRO_ID ipat
     | _ -> tclUNIT () in
-  let unprot =
+  let unprotect =
     if eqid <> None && is_rec
     then V82.tactic ~nf_evars:false Ssrcommon.unprotecttac else tclUNIT () in
-  V82.of_tactic begin
-    V82.tactic ~nf_evars:false ssrelim <*>
-    tclIPAT_EQ (intro_eq <*> unprot) ipats
-  end
+  begin match seed with
+  | None -> ssrelim
+  | Some s -> IpatMachine.tclSEED_SUBGOALS s ssrelim end <*>
+  tclIPAT_EQ (intro_eq <*> unprotect) ipats
+;;
 
 let mkEq dir cl c t n env sigma =
   let open EConstr in
@@ -506,13 +670,11 @@ let ssrelimtac (view, (eqid, (dgens, ipats))) =
     | [v] ->
       Ssrcommon.tclINTERP_AST_CLOSURE_TERM_AS_CONSTR v >>= fun cs ->
       tclDISPATCH (List.map (fun elim ->
-        V82.tactic ~nf_evars:false
           (Ssrelim.ssrelim deps (`EGen gen) ~elim eqid (elim_intro_tac ipats)))
         cs)
     | [] ->
       tclINDEPENDENT
-        (V82.tactic ~nf_evars:false
-          (Ssrelim.ssrelim deps (`EGen gen) eqid (elim_intro_tac ipats)))
+          (Ssrelim.ssrelim deps (`EGen gen) eqid (elim_intro_tac ipats))
     | _ ->
       Ssrcommon.errorstrm
         Pp.(str "elim: only one elimination lemma can be provided")
@@ -535,9 +697,8 @@ let ssrcasetac (view, (eqid, (dgens, ipats))) =
             if view <> [] && eqid <> None && deps = []
             then [gen], [], None
             else deps, clear, occ in
-          V82.tactic ~nf_evars:false
-            (Ssrelim.ssrelim ~is_case:true deps (`EConstr (clear, occ, vc))
-              eqid (elim_intro_tac ipats))
+          Ssrelim.ssrelim ~is_case:true deps (`EConstr (clear, occ, vc))
+            eqid (elim_intro_tac ipats)
       in
       if view = [] then conclusion false c clear c
       else tacVIEW_THEN_GRAB ~simple_types:false view ~conclusion info)
@@ -556,18 +717,16 @@ let pushmoveeqtac cl c = Goal.enter begin fun g ->
   Tactics.apply_type ~typecheck:true (EConstr.mkProd (x, t, cl2)) [c; eqc]
 end
 
-let eqmovetac _ gen = Goal.enter begin fun g ->
-  Ssrcommon.tacSIGMA >>= fun gl ->
-  let cl, c, _, gl = Ssrcommon.pf_interp_gen gl false gen in
-  Unsafe.tclEVARS (Tacmach.project gl) <*>
-  pushmoveeqtac cl c
-end
+let eqmovetac _ gen =
+  Ssrcommon.pfLIFT (Ssrcommon.pf_interp_gen false gen) >>=
+  fun (cl, c, _) -> pushmoveeqtac cl c
+;;
 
 let rec eqmoveipats eqpat = function
   | (IPatSimpl _ | IPatClear _ as ipat) :: ipats ->
        ipat :: eqmoveipats eqpat ipats
   | (IPatAnon All :: _ | []) as ipats ->
-       IPatAnon One :: eqpat :: ipats
+       IPatAnon (One None) :: eqpat :: ipats
   | ipat :: ipats ->
        ipat :: eqpat :: ipats
 
@@ -700,8 +859,8 @@ let ssrabstract dgens =
      let open Ssrmatching in
      let ipats = List.map (fun (_,cp) ->
        match id_of_pattern (interp_cpattern gl0 cp None) with
-       | None -> IPatAnon One
-       | Some id -> IPatId id)
+       | None -> IPatAnon (One None)
+       | Some id -> IPatId (None,id))
        (List.tl gens) in
      conclusion ipats
   end in